fix: addressed all PR comments
