+++ src/utils/resources/udns/constants.ts
@@ -5,3 +5,5 @@ export const FIXED_PRIMARY_UDN_NAME = 'primary-udn';
 export const LOCALNET_TOPOLOGY = 'Localnet';
 export const LAYER2_TOPOLOGY = 'Layer2';
 export const LAYER3_TOPOLOGY = 'Layer3';
+
+export const CLUSTER_UDN_NAD_NAME_PREFIX = 'cluster.udn.';
