+++ src/utils/resources/vm/constants.ts
@@ -1,2 +1,3 @@
 export const NETWORK_PATH = '/spec/template/spec/networks';
 export const INTERFACE_PATH = '/spec/template/spec/domain/devices/interfaces';
+export const INTERFACE_STATE_ABSENT = 'absent';
