+++ src/views/nads/details/hooks/useNADVirtualMachines/index.ts
@@ -1,0 +1,1 @@
+export { default, type NADVirtualMachine } from './useNADVirtualMachines';
