+++ src/views/nads/details/tabs/details/NADDetails.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import * as _ from 'lodash';
+import { size } from 'lodash';
 
 import { modelToGroupVersionKind } from '@kubevirt-ui/kubevirt-api/console';
 import NetworkAttachmentDefinitionModel from '@kubevirt-ui/kubevirt-api/console/models/NetworkAttachmentDefinitionModel';
@@ -26,6 +26,7 @@ import Loading from '@utils/components/Loading/Loading';
 import { OwnerReferences } from '@utils/components/OwnerReference/owner-references';
 import { useNetworkingTranslation } from '@utils/hooks/useNetworkingTranslation';
 import { NetworkAttachmentDefinitionKind } from '@utils/resources/nads/types';
+import { getAnnotations, getLabels, getName, getNamespace, getUID } from '@utils/resources/shared';
 
 import NADDescriptionDetails from './NADDescriptionDetails';
 import NADTypeDetails from './NADTypeDetails';
@@ -36,14 +37,15 @@ type NADDetailsProps = {
 
 const NADDetails: FC<NADDetailsProps> = ({ obj: nad }) => {
   const { t } = useNetworkingTranslation();
-  const metadata = nad?.metadata;
+  const name = getName(nad);
+  const namespace = getNamespace(nad);
   const annotationsModalLauncher = useAnnotationsModal(nad);
   const labelsModalLauncher = useLabelsModal(nad);
 
   const [canUpdate] = useAccessReview({
     group: NetworkAttachmentDefinitionModel.apiGroup,
-    name: metadata?.name,
-    namespace: metadata?.namespace,
+    name,
+    namespace,
     resource: NetworkAttachmentDefinitionModel.plural,
     verb: 'patch',
   });
@@ -56,6 +58,8 @@ const NADDetails: FC<NADDetailsProps> = ({ obj: nad }) => {
     );
   }
 
+  const annotationCount = size(getAnnotations(nad, {}));
+
   return (
     <PageSection>
       <DetailsSectionTitle
@@ -65,13 +69,13 @@ const NADDetails: FC<NADDetailsProps> = ({ obj: nad }) => {
         <GridItem md={6}>
           <DL className="co-m-pane__details" data-test-id="resource-summary">
             <DetailsItem label={t('Name')} obj={nad} path="metadata.name" />
-            {metadata?.namespace && (
+            {namespace && (
               <DetailsItem label={t('Namespace')} obj={nad} path="metadata.namespace">
                 <ResourceLink
                   kind="Namespace"
-                  name={metadata.namespace}
+                  name={namespace}
                   namespace={null}
-                  title={metadata.uid}
+                  title={getUID(nad)}
                 />
               </DetailsItem>
             )}
@@ -86,7 +90,7 @@ const NADDetails: FC<NADDetailsProps> = ({ obj: nad }) => {
             >
               <LabelList
                 groupVersionKind={modelToGroupVersionKind(NetworkAttachmentDefinitionModel)}
-                labels={metadata?.labels}
+                labels={getLabels(nad)}
               />
             </DetailsItem>
             <DetailsItem label={t('Annotations')} obj={nad} path="metadata.annotations">
@@ -100,12 +104,12 @@ const NADDetails: FC<NADDetailsProps> = ({ obj: nad }) => {
                   variant={ButtonVariant.link}
                 >
                   {t('{{count}} annotation', {
-                    count: _.size(metadata?.annotations),
+                    count: annotationCount,
                   })}
                 </Button>
               ) : (
                 t('{{count}} annotation', {
-                  count: _.size(metadata?.annotations),
+                  count: annotationCount,
                 })
               )}
             </DetailsItem>
@@ -113,7 +117,7 @@ const NADDetails: FC<NADDetailsProps> = ({ obj: nad }) => {
               <NADDescriptionDetails nad={nad} />
             </DetailsItem>
             <DetailsItem label={t('Created at')} obj={nad} path="metadata.creationTimestamp">
-              <Timestamp timestamp={metadata?.creationTimestamp} />
+              <Timestamp timestamp={nad.metadata?.creationTimestamp} />
             </DetailsItem>
             <DetailsItem label={t('Owner')} obj={nad} path="metadata.ownerReferences">
               <OwnerReferences resource={nad} />
