+++ src/views/nads/details/tabs/virtualMachines/utils.ts
@@ -1,0 +1,15 @@
+export const getVMStatusLabelColor = (status: string) => {
+  if (!status) {
+    return 'grey';
+  }
+
+  if (status.includes('Error') || status === 'Stopped') {
+    return 'red';
+  }
+
+  if (status === 'Running') {
+    return 'green';
+  }
+
+  return 'orange';
+};
