+++ src/views/udns/details/hooks/useUDNVirtualMachines/index.ts
@@ -1,0 +1,1 @@
+export { default, type UDNVirtualMachine } from './useUDNVirtualMachines';
