+++ src/views/udns/details/tabs/virtualMachines/constants.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_PER_PAGE = 20;
+
+export const ADD_VM_MODAL_COLUMN_COUNT = 4;
+export const ADD_VM_MODAL_SORT_COLUMN_KEYS = ['name', 'namespace', 'status'] as const;
